Use travis for uploading coverage reports to coveralls
currently Github actions do not expose secrets for PRs from
forks. This means that in order to upload coverage reports for PRs
from a github action we would have to expose the REPO_TOKEN. As a work
around we can build on travis for PRs and upload to coveralls from
there.