Implementation of the QCRad algorithm for irradiance quality control (#19)
This implementation is based on code from solarforecastarbiter (SFA) with two major changes:

1. We remove the @mask_flags decorator from the filtering function. This decorator is specific to SFA's bit-wise masks and not needed for this general purpose library.
2. We do not set the Series.name field of the returned masks. In a general purpose library it is not clear what the correct name should be, so for now we will leave it to the user to set the name if they need it.

* Add implementation of QCRad quality checks to quality.irradiance

* Adds tests directory and test module to cover the new irradiance checks.

* Adds a licenses.rst file to the documentation to keep track of
  open source licences and copyright notices.

* Add API documentation

Enables the autodoc, autosummary, and napoleon sphinx extensions to
generate docs from docstrings.

Adds api file to documentation tree with documentation for the qcrad
functions in `quality.irradiance`

* Docstring changes

- Properly format math expressions.
- Remove references from summary lines.
- Clean up return types for check_irradiance_limits_qcrad
- Format notes/warnings with reST directives

* Use raw strings for docstrings with math

These docstrings contian a '\' character (escapes the _ in the
generated docs) that should be interpreted literally.

* Add test for the internal `_check_limits` function

The qcrad checks do not fully exercise this function on their own.

* Break lines before binary operators

PEP 8 recommends line breaks before binary operators rather than
after.

* Follow numpydoc rules for refering to parameters

* Follow PEP 257 guidelines for doc strings.

- Use the imperative voice in first line ("test for..." rather than
  "tests for...").
- Single line doc strings fit on one line (including quotes)
- No blank line after function doc string.

* Limit the description of the limits dict to only relevant keys

For check_*_limits_qcrad functions limits only needs to have keys
corresponding to the specific measurement being
checked (ghi/dhi/dni). This change makes it clear what keys are
actually required for each function. Additional keys are
allowed (i.e. we use QCRAD_LIMITS as the default for all functions)
but are ignored.

* Do not set the name of the series returned by qcrad functions

The name of the series was used in solarforecastarbiter, but it does
not add value here.

Because the name is not set, we also have to update the tests to
ignore `Series.name`.

* Note type of dict values in doc for check_irradiance_limits_qcrad

* Remove name of return value from docstrings for functions that do not return tuples

Including the name doesn't add any information for functions that
return only one value.

Co-authored-by: Cliff Hansen <cwhanse@sandia.gov>