Port interpolation and stale value detection from SolarForecastArbiter (#22)
* Add interpolation & stale value filters from SolarForecastArbiter

- removes the @mask_flags decorator and updates docstrings
- rename detect_stale_values to stale_values_diff
- rename detect_interpolation to interpolation_diff

* Move stale data and interpolation into the quality module

Create a new gaps submodule in quality for functions that identify
gaps/missing data.

Update README to describe the quality.gaps module.

* Add documentation for quality.gaps module

* Add reference to gaps functions in licenses documentation

These two functions are derived from SFA, we need to properly
attribute them in the documentation.

* Names of gaps filtering funcitons reflect their implementation

In the future we will add other methods for detecting stale
values/interpolation. In the style of pvlib this commit adds a
trailing '_diff' to the function names (the detection is based on the
difference between adjacent data).