Port detect_clipping from Solar Forecast Arbiter (#31)
* validator.detect_clipping is renamed to features.clipping_levels

In the future we will add different methods for detecting clipping. This naming
convention matches the convention in pvanalytics.quality as will as in
pvlib-python.

* simplify tests by removing pvlib dependency

Reimplements a simpler/pared down version of the tests from SFA using
a quadratic for data rather than generating GHI data using
pvlib.location.Location.get_clearsky(). Using pvlib to generate testing data 
significantly complicates our CI process because it requires libraries that
are not available on the GitHub runners (libhdf5).

* Add parameter defaults from SFA

* Add more complex tests for clipping

The idea is to test data that looks like the summed output of two
inverters where one inverter is experiencing clipping.

* Add clipping_levels to API documentation

Also adds attribution in licenses.rst

* Apply documentation changes suggested in code review

Co-Authored-By: Cliff Hansen <cwhanse@sandia.gov>

* Test data drops below the clipping level in the middle of clipping

This give two periods of clipping surrounding a period without
clipping.

* Docstring for 'quadratic' fixture matches index of returned series.

The quadratic describeds in the previous docstring is actually scaled
and shifted into the index of the series which is the integers [0, 59].

* Reindex quadratic test fixture so vertex falls at index 30

Previously the vertex was between 30 and 31, this change should make the fixture easier to work with going forward.

Co-authored-by: Cliff Hansen <cwhanse@sandia.gov>