Add copyright notice to each function ported from SolarForecastArbiter (#34)
* Add copyright notice to each function ported from SolarForecastArbiter

Finer-grained copyright notices/tracking will be helpful as we
incorporate functionality from other projects. Keeping the copyright
in the individual function's docstring should make tracking/updating
copyright notices simpler in the future.

* Create directory for third party licenses

We will add a copy of the license for each third party project we pull
code from to the LICENSES directory. For each function that is ported
we specify the copyright and reference the license for the project
from which the function was ported.

* Remove third party licenses page from documentation.

These licenses are now included in the LICENSES directory and the
appropriate copyright attribution is included in the documentation for
each function.