Relax the criteria for excluding night time data.
By keeping more data we can handle the case where there is the same
number of positive power data points at each minute. With a strictly
greater than comparison we end up excluding all the data in this
case. Including data at the first quartile shouldn't have a
substantial effect for large datasets with night time data, and it
makes the filter more useful for a variety of data sets.