First-cut of function to identify daytime
Not working yet, but includes a couple initial tests and the rough
documentation of the function.

'features.time' may not be the right place for it, but it makes sense
to me at the moment. I think this kind of function can be useful in a
quality check, but in and of itself it is not a QA function---it is a
feature extraction function. Another organization option could be
'features.irradiance' (since the function operates on irradiance
data); however, because it can also operate power data it doesn't make
sense to me to put it there.