Additional single-day tests for time.daytime
Fixes bug where maximum number of days with data was not calculated
correctly. Since the data was transformed from a series to a frame the
return value of 'data.max()' was a series. The conditional, however,
was expecting a scalar. To avoid this we just compute the minutes of
the day as their own series and use that series to do the grouping
rather than making the minutes of the day another column of the data.