Refactor _clipping_power to avoid looping over the data
Makes the calculation of the clipping power level a bit more readable,
it is possible that it also improves the performance by using pandas
functions instead of loops, but I doubt the difference is noticable
given that the loop was only iterating over a single day of data (at
most 1440 values). Readability is substantially better, though.