First cut add curve fitting and orientation inference logic
Trying to take just the core functionality form PVFleets QA.  These
are the main differences:

1. Data is not separated into winter and summer months. This means
that if there are time-shifts in the data they could mess up the
fit. By removing the separation we get the core
functionality (curve fitting) and can leave it to the user to
pass only winter/summer and combine the orientation inferred for each
in to a meaningful result (i.e. PVFleets basically decides the
orientation is inconclusive if the winter and summer don't match)

2. A Boolean mask for day/night is passed as a parameter

3. Clipping is passed as a Boolean mask, and the percent of clipping
is calculated (rather than passing in clip_percent directly)

4. The responsibilities of the function have been made narrower,
reducing the number of values returned and hopefully making future
maintenance simpler. It is possible (probable?) that I went too far in
removing responsibilities and that some additional return values are
needed. When this comes up we should look carefully at whether to add
them here or add new functions with separate responsibilities.

Going forward I need to make a testing plan for this and write some
tests. Also need to decide whether returning None is a good idea, or
if I should add a third instance to the Orientation
Enum (e.g. INDETERMINANT).

Finally I would like to do a bit of simplification in the
_orientation_from_fit function. It would be nice if this could be made
more general/flexible (i.e. pass in the various thresholds), it will
take a bit of exploring to figure out the best way forward for this.