Rough first cut laying out scaffolding for daily orientation checks
Need to add curve fitting functions to an internal module that can be
used by other parts of pvanalytics (system.py, for example). Will
probably find other things down the line to put in that module. Also
includes a start on the labeling code which will resample by day, and
apply an _is_sunny() predicate to determine whether the curve fit was
good. Then we can apply the usual tricks for expanding those truth
values out to all the timestamps on 'sunny' days.