Documentation edits from code review
Change up the order of sentences in the last paragraph (much better - thanks, Cliff). rewording in the third paragraph.

Need to find a way to describe that only the values for the longest period with low-slope high-power are used to compute the threshold. (as is, it reads kind of like we take the average of the full curve).

Co-authored-by: Cliff Hansen <cwhanse@sandia.gov>