Rough implementation of the separate tracking and fixed function
Adds the pre-filtering steps for low power from the original PVFleets
code, need to figure out the tests that are failing now. Because of
the extra filtering step the perturbed days (which have just a
constant value in the tests) are now eliminated and are not set to
False in the output from either function. Might be able to fix it by
setting a fill value in the reindex operation.