First working implementation of tracking/fixed detection
Change how data is grouped/resampled to work around pandas bug (issue
is #25411). This bug causes curve fits to fail because data gets
shifted across midnight.

Abstracts the pattern for fitting curves only if some predicate is
satisfied. If the predicate is not satisfied then the fit is assigned
an r^2 of 0.0.