Expaning overview of features.orientation functions in API docs
Rewrote this in an attempt to capture the limitations and use cases
for these two functions. The most important thing to capture is that
the negation of the fixed mask is not equivalent to days when the
tracker is not working. Suggest combining the two funcitons to get at
days where the tracker is stuck.