Rename system.orientation and other related elements
Orientation is misleading (makes you think azimuth and tilt). This
renames system.orientation() to system.is_tracking_envelope() and
system.Orientation to system.Tracker. The function name is much
better, capturing the semantics well and keeping space in the library
for other methods to be added in the future. The name Tracker for the
enum should be generic enough that more specific types of trackers
could be added to it (for example if future functions support
differentiatin between single- and dual-axis trackers).