Guard against empty winter/summer data
This is maybe not perfect, but it gets the main point across. If one
or both of the summer/winter subsets are empty then just do the curve
fitting on the full data set. We may want to change this criteria to
fit the full series of there is less than X amount of data in one
half, but this will work for now.