Don't resample dataframe, capture 'hours' in closure environment
Different versions of Pandas seem to have different bahavior when
resampling a DataFrame. Newer versions (> 1.0) pass a Series to the
function in `apply()`, dropping all other columns from the
DataFrame. This means that the function cannot operate on more than
one column at a time. To work around that we create a new series for
the hour-of-day (the x-values for integration) and cature it in the
closure environment. I think this should perform well enough, this is
also only for the 'slow case' where the timestamp spacing is
irregular.