Find solar noon directly from maximum clearsky POA
When identifying the tilt and azimuth in `system.orientation()` we
calculate solar noon on each day by fitting a quadratic to the power
or irradiance measurements and comparing this to solar noon for
clearsky POA irradiance at different orientations. Rather than fitting
a quadratic to the daily clearsky POA irradiance and using the vertex
of the quadratic as solar noon this commit uses the maximum POA value
for the day. The idea is that curve fitting is unnecessary---because
clearsky POA has no noise or interference we can identify solar noon
directly from the data. This matches the methodology in PVFleets QA
and should improve performance; however, using the maximum POA
directly can reduce the accuracy of the tilt & azimuth that are
returned if the frequency of clearsky data is low. In general the best
accuracy can be obtained by passing one-minute clearsky and solar
position data.

Because of the reduced accuracy with low-frequency clearsky data the
unit tests needed to be updated. Rather than extending the range of
'okay' tilt/azimuth values I elected to pass one minute clearsky. For
one-minute data we know that the accuracy should be +/- 10 degrees.