Daily insolation limits (#57)
Check that the insolation for each day is within a reasonable range
of clearsky insolation for each day.

* Use scipy.integrate to compute daily insolation with trapezoid rule

Uses two different approaches depending on whether the series has
regular timestamp spacing or not. With regular timestamp spacing
we pass `dx=<freq-in-hours>` directly to the integration function. If 
timestamps are not evenly spaced then the time of day in hours is 
calculated for each value and used as the x-coordinates in the 
integration.

Co-authored-by: Cliff Hansen <cwhanse@sandia.gov>