Test correlation between module temperature and irradiance (#55)
Add a pass/fail quality check for PV module temperature
data. Unlike other functions in this module, the entire data set
is considered, and passed or failed as a whole. In the future we
may want to check for correlation on a rolling basis to identify
specific periods of time where the sensor is malfunctioning.

Co-authored-by: Cliff Hansen <cwhanse@sandia.gov>