Refactor curve fitting functions (#63)
Since these functions used in a pandas.GroupBy.apply() call,
manipulating the index to calculate time of day in minutes can
significantly reduce performance. Here we remove the
_get_minute_of_day() function and instead pass in the x values as a
paramter. In addition to improving performance this makes the curve
fitting code cleaner and substantially more general.

* Add examples for using quadratic and quartic_restricted

Shows how to calculate numeric x-values for curve fitting to a time
series.

* Expand comment describing operation of _conditional_fit() function

Bring the comment up to date and provide a description of all
parameters. This function is sufficiently complex to require a
docstring-like comment to describe it.

Co-authored-by: Cliff Hansen <cwhanse@sandia.gov>