Make _fit.quadratic_idxmax API consistent with other _fit functions
All API functions in _fit have a consistent interface, taking
array-like x-values and a Series of y-values. Previously we took two
series and indexed into the x-values using the index of the
y-values. Now, we expect users to pass arrays/series of matching
length, one with x-value and one with y-values. This shifts the
use-case specific code (i.e. converting dates) out of the general
purpose curve fitting functions an in to the code that calls the
fitting function. I take one more step to making this as generic as
possible by adding a `model_range` parameter, used to pass the range
of x-values to apply the quadratic to after it is fit. The previous
fitting functions did not need this parameter since they just return
the r-squared for the fit.