Expect higher irradiance during the day
Non-zero irradiance is ont a reasonable criteria for daytime. Using
this could lead to including values like 0.5 W/m^2 as day which is not
something we should expect the daylight filtering functions to be able
to detect. 2 W/m^2 is still an arbitrary threshold, but it at least
prevents the situation above.