Reduce the number of candidate orientations in tests
We can get away with testing a smaller number of candidate tilts and
azimuths in the unit tests for this function. We sacrifice testing
that the accuracy is as good as we want it to be, but still test that
the funciton works more or less as expected. This change reduces the
running time of the test suite by 10 seconds, so I think it is worth
it.

Since we are using fewer candidates, I removed the _assert_within()
function as it is no longer needed.