Correct errors in the day/night mask where night is too long
These errors arise when input data goes to zero (or low) shortly
before sunset and remains there in to night. After the midday errors
are corrected by _filter_midday_errors() and clipping is flagged as
day we look at the two-week rolling median of length of each daytime
period, flagging days where the daytime period is more than 30 minutes
shorter than the median. On these days we replace the data with the
by-minute 31-day rolling median in the same way as
_filter_midday_errors().

Expanded test cases to cover both sunrise and sunset errors as well as
short and long errors.