Handle missing data in daytime.diff
Add parameter to specify frequency for Series where the frequency
cannot be inferred.

NaNs are treated like zeros. I'm not 100% sure this is the right
decision. In cases where there are many days of NaNs (more than a week
or two) it could cause problems with the good data to either
side. That said, getting a not-perfect mask for just the edges of a
gap is probably okay. In fact, I like this way better because it lets
you use the day/night mask based on a faulty irradiance sensor to
select daytime in some other data, which may not have the same gaps as
the irradiance data.