Rename infer_orientation_solarnoon to infer_orientation_daily_peak
Solar noon has a different meaning than what we are doing here. It
means the time mid-way between sunrise and sunset and may not
necessaryily be the same time as peak poa irradiance or power. Since
we are actually looking at peak power/poa, we are changing the name to
reflect that.