Infer azimuth and tilt from power or POA irradiance (#64)
* Add system.infer_orientation_daily_peak() function

Implement the tilt & azimuth inference algorithm used in the
PVFleets quality assurance project. This algorithm works by
fitting a quadratic to the power or POA irradiance for each
day. The solar azimuth coinciding with the vertex of the
quadratic is compared to solar azimuth at maximum POA at several
candidate orientations and the best fit is returned.

* Add _fit.quadratic_vertex() function

Adds a new function to the internal curve fitting API that
returns the x-value at the vertex of a quadratic fit to the data.

* Rename curve fitting functions

Rename _fit.quadratic() to _fit.quadratic_r2() and
_fit.quartic_restricted() to _fit.quartic_restricted_r2() to make
it more clear what these functions do.

Co-authored-by: Cliff Hansen <cwhanse@sandia.gov>