Determine whether a system has a fixed orientation or is equipped with a tracker (#49)
* Use an enum to specify whether a system is equipped with a tracker

Add system.Tracker enum with instances FIXED, TRACKING, and UNKNOWN

* Add test for GHI orientation.

The orientation of a GHI sensor is FIXED.

* Add orientation test based on simulated power data

Uses pvlib to simulate power from a fixed PV system and verifies that
system.is_tracking_envelope() returns Tracker.FIXED.

* Test that a simulated single axis tracker is Tracker.TRACKING

* data with too much clipping has UNKNOWN orientation

* Test that a constant signal has "UNKNOWN" orientation

* Test orientation on tracking system that is broken part of the time

If the profile of the median does not match the profile of the 99.5%
quantile (because 100 days have garbage data) then the orientation is
UNKNOWN.

* Suppress warning caused by test data with slope equal to zero

Zero slope (in the constant data) leads to warnings from the curve
fitting funcitons due to division by zero.

* Test that a fixed system with perturbed data has UNKNOWN orientation

100 days of garbage data messes up the median and causes the fit to
fail when 'fit_median' is True.

* Added system.is_tracking_envelope to API documentation

* Remove empty test.util.test_fit module

No longer need this test since the interface for the curve fitting
functions now accepts non-datetime indexed series.

* Add parameter to control minimum value of daytime data

Trying to propagate magic-number type values up to the API. These
parameters could be important for tuning performance on particular
data sets.

* Rewrite is_tracking_envelope() docstring

Reword and reformat to clarify the conditions that must be met for a system to be marked as TRACKING, FIXED, or UNKNOWN. Also make clear the effect of the optional median fit.

Co-authored-by: Cliff Hansen <cwhanse@sandia.gov>

* Move maximum clipping percent test to top-level function

Instead of passing clipping percentage and maximum clipping percent
down through three layers of internal functions we perform the
clipping percent test and get the fit bounds at the top level. This
makes the code somewhat more efficient by short-circuiting the
resampling/grouping steps if there is too much clipping. The number of
parameters that are threaded through the internal functions is also
reduced by two, which should improve maintainability and readability.

* Use clipping fraction instead of percent

Use the ratio of clipped measurements to the total number of
measurements directly rather than converting to a percent. With a
fraction between 0 and 1 the fit params dicitonary is a little easier
to understand.

* Make seasonal_split a dictionary

Change the semantics of the seasonal_split param. Now expects
a dictionary with keys 'winter' and 'summer'. The behavior of
`infer_tracking_envelope()` is also updated so if either season
has no data, curve fitting proceeds only on the season that has
data (previously the entire series was used).

* Issue a warning when there is no data for both winter and summer

We just return UNKNOWN in this case; however, it is possible that
a user made a mistake in this case. Providing a warning seems like
the polite thing to do.

* Reword documentation

Highlight that seasonal grouping is optional and improve
description of `seasonal_split` parameter.

Co-authored-by: Cliff Hansen <cwhanse@sandia.gov>