Use majority truth-value for error correction
Rather than converting the series to a numeric representation and using
the median for error correction we use the majority truth-value. It is
not necessary to do the numeric conversion since any NAs in the series
will be automatically converted to False in the comparisons that precede
the correction operation(s).