Identify when the sun is up based on power or irradiance (#67)
* Implementation of the PVFleets day/night detection algorithm

Add the function features.daytime.power_or_irradiance() to identify
day and night from power or irradiance data.

* Test only one day in the series with zeros at mid-day

It will not work if every day has the same hours zeroed, there is no
information at that point to detect that the zeroes are an anomaly.

* Ignore low-ghi times in daytime tests

I don't have a good way of knowing exactly how accurate the daytime
mask should be. Mainly what we want to test is whether the values
where clearsky GHI is 0 are flagged as night and the values where
clearsky GHI is large are flagged as day.

* Add daytime mask test with different timestamp spacings

Adds tests at 1-hour spacing and at 1-minute spacing.

* Handle arbitrary timestamp spacing

Adjust the derivative threshold and the minimum run lengths to account
for input with arbitrary, but uniform, timestamp spacing.

* Add test covering daylight savings shifts

Seems like a good idea to cover these time periods when dealing with
day/night filters. Required a change to the shoulder value, relaxing
the criteria slightly, but I think 3 W/m^2 is still strict enough.

* Test daytime edge case and missing data

- Test the edge case where data goes to 0 mid-day and stays zero
  throughout the rest of the day (low value/low derivative runs in to
  night time)

- Test missing data where missing data is marked with NaN and missing
  data is completely missing

* Correct errors in the day/night mask where night is too long

These errors arise when input data goes to zero (or low) shortly
before sunset and remains there in to night. After the midday errors
are corrected by _filter_midday_errors() and clipping is flagged as
day we look at the two-week rolling median of length of each daytime
period, flagging days where the daytime period is more than 30 minutes
shorter than the median. On these days we replace the data with the
by-minute 31-day rolling median in the same way as
_filter_midday_errors().

Expanded test cases to cover both sunrise and sunset errors as well as
short and long errors.

* Refactor test_daytime.py

Use pytest parameters to run every test for different timestamp
spacings (1-minute, 15-minute, and 1-hour). Consolidates tests and
removes duplicated setup code.

* Skip slow tests by default

Mark daytime tests with 1-minute clearsky data as slow and skip them
by default. Adds the '--runslow' option to tell pytest not to skip
them.

* Add --runslow option to CI pipeline

* Change tests to demand less precision for 1-minute timestamp spacing

Not reasonable to require the algorithm to determine sunrise/sunset
within one-minute when the beginning/end of the day is missing data.

* Handle missing data

Add parameter to specify frequency for Series where the frequency
cannot be inferred.

NaNs are treated like zeros. I'm not 100% sure this is the right
decision. In cases where there are many days of NaNs (more than a week
or two) it could cause problems with the good data to either
side. That said, getting a not-perfect mask for just the edges of a
gap is probably okay. In fact, I like this way better because it lets
you use the day/night mask based on a faulty irradiance sensor to
select daytime in some other data, which may not have the same gaps as
the irradiance data.

* Test with outliers

Pass a Boolean series indicating that some values are outliers. The
returned mask should mark the values correctly despite being outliers.

* Documentation for dayime mask function

Add overview of algorithm to the docstring and add the API function to
the API docs.

* Fix typo in comment

* Limit pandas to below version 1.1.0

Work around pandas bug in 1.1.x for groupby().rolling(center=True)

* Add test with simulated variable days

* Describe low_median_threshold as a 'maximum'

All the thresholds are applied as maximums.

* Rename daytime.diff daytime.power_or_irradiance

New name reflects that the function operates on power/irradiance data
as opposed to solar position or some other input.

* Move 'albuquerque' fixture to conftest

I keep re-defining this fixture. Move it to conftest so it is
available wherever it is needed.

* Pass parameters for controlling error correction

Several parameters can be tuned for the error correction steps:
- window used for smoothing/correction
- window used for getting the median day length
- maximum acceptable difference between length of day and the median
  length of day in surrounding X days
- minimum number of hours for day and night periods

* Make window for calculating the median value at each minute a param

* Reword documentation

Substantial improvements from code review.

Co-authored-by: Cliff Hansen <cwhanse@sandia.gov>

* Use majority truth-value for error correction

Rather than converting the series to a numeric representation and using
the median for error correction we use the majority truth-value. It is
not necessary to do the numeric conversion since any NAs in the series
will be automatically converted to False in the comparisons that precede
the correction operation(s).

Co-authored-by: Cliff Hansen <cwhanse@sandia.gov>