Exclude the change point from shift labels
The index returned by ruptures is the first index of the changed
signal. We change the call to pandas.cut() to be closed on the
left and open on the right so that the shift labels are assigned
correctly.

Also pass jump=1 to ruptures.Pelt.__init__ to get the best
precision in the changepoint detection results. We may want to
allow users to pass a different `jump` if they want to get better
performance at the price of reduced precision.