Only append last index if not identified as a breakpoint
Ruptures seems to always return the last index when using Pelt;
however, there does not appear to be any guarantee of this in the
documentation. To make sure that the entire series is covered by
the intervals passed to cut we add the last index+1 if it is not
already present in the breakpoint list.