Tweak the ruptures.Pelt.predict() penalty parameter
Reducing the penalty slightly gets all the tests passing. I am
slightly concerned about whether the value we are using (13) is
actually a reasonable value. The effect of `pen` and its interaction
with the `min_size` parameter to ruptures.Pelt() makes me think
that `pen` (and other ruptures-related paramters) should be exposed
in the API.