Explicitly unset the frequency for the daytime mask in tests
Having a frequency set creates problems in later versions of pandas
that add the `check_freq` param to assert_series_equal. We do get
the same index that was passed in, but some of the tests mess up
the index in ways that are difficult to capture. Better to just
drop the freq all together.