Time shift detection (#74)
* Implementation of ruptures-based time shift detection

* Add optional dependency on ruptures 

* Add tests for +/- 1-hour shifts

Create a fixture that can be used for most/all time-shift tests.

* Exclude the change point from shift labels

The index returned by ruptures is the first index of the changed
signal. We change the call to pandas.cut() to be closed on the
left and open on the right so that the shift labels are assigned
correctly.

Also pass jump=1 to ruptures.Pelt.__init__ to get the best
precision in the changepoint detection results. We may want to
allow users to pass a different `jump` if they want to get better
performance at the price of reduced precision.

* Only append last index if not identified as a breakpoint

Ruptures seems to always return the last index when using Pelt;
however, there does not appear to be any guarantee of this in the
documentation. To make sure that the entire series is covered by
the intervals passed to cut we add the last index+1 if it is not
already present in the breakpoint list.

* Refactor rounding to support more conservative shift estimates

Adds parameters to specify where to round to and the point at
which to round up instead of down. Will facilitate the more
conservative shift estimation in the PVFleets project, just need
to expose these two params at the top-level API.

* Tweak the ruptures.Pelt.predict() penalty parameter

Reducing the penalty slightly gets all the tests passing. I am
slightly concerned about whether the value we are using (13) is
actually a reasonable value. The effect of `pen` and its interaction
with the `min_size` parameter to ruptures.Pelt() makes me think
that `pen` (and other ruptures-related paramters) should be exposed
in the API.

* Document the algorithm implemented by time.shifts_ruptures()

* Explicitly unset the frequency for the daytime mask in tests

Having a frequency set creates problems in later versions of pandas
that add the `check_freq` param to assert_series_equal. We do get
the same index that was passed in, but some of the tests mess up
the index in ways that are difficult to capture. Better to just
drop the freq all together.

* Add quality.time.shifts_ruptures to API documentation

* Return a Boolean mask as well as shift amounts

Return a Boolean mask that can be used to determine whether shifts exist
without caring about the actual shift amount. The shift amount is also
returned as a separate series.

* Add notes on how to use time.shifts_ruptures()

Co-authored-by: Cliff Hansen <cwhanse@sandia.gov>