Fix MAD scaling using of the `c` parameter to robust.scale.mad()
statsmodels.robust.scale.mad() handles scaling internally. In fact the
previous implementation was incorrect, relying on the raw MAD rather
than scaling it by the `scale` parameter in the default usage, and
scaling incorrectly if users pass a non-default `scale`. This commit
fixes the usage of mad() and fixes a test which was also implemented
incorrectly.

The changes to the test change the `maximum_deviation` cutoff and now
reflect the true variability in the data. Outliers are placed at +/-
5, 10, and 15 times the variance. We now pass these values as the
cutoff in each test and see the correct outliers allowed to pass.