Reduce duplicate code for converting freqstr to Timedelta (#84)
* Add function for converting freqstr to Timedelta to `pvanalytics.util`

Adds a new utility function that is used anywhere we converted a
frequency string to a Timedelta. Previously there was substantial
code duplication in these places. There are still some other functions
that already use the tseries API that we should refactor to use the
new utility function.

* Replace direct use of frequencies API with util.freq_to_timedelta()

This takes care of the existing usage of the tseries API mentioned
above.