Use statsmodels for maximum absolute deviation (#85)
* Use statsmodels for maximum absolute deviation

Rather than scipy.stats we will use statsmodels.robust.scale.mad()
to calculate the Maximum Absolute Deviation. This change will let us
lower the minimum scipy version to 1.2
(scipy.stats.maximum_absolute_deviation was introduces in 1.3).

* Lower minimum scipy version to 1.2

* MAD scaling using of the `c` parameter to robust.scale.mad()

statsmodels.robust.scale.mad() handles scaling internally. In fact the
previous implementation was incorrect, relying on the raw MAD rather
than scaling it by the `scale` parameter in the default usage, and
scaling incorrectly if users pass a non-default `scale`. This commit
fixes the usage of mad() and fixes a test which was also implemented
incorrectly.

The changes to the test change the `maximum_deviation` cutoff and now
reflect the true variability in the data. Outliers are placed at +/-
5, 10, and 25 times the variance. We now pass these values as the
cutoff in each test and see the correct outliers allowed to pass.

* Default to scale=None in pvanalytics.outliers.hampel()

Use None to specify the more precise value used by default in
the statsmodels function. Add test that verifies the scale parameter
has an effect (does not verify much beyond this; relies on the
correctness of the statsmodel implementation).