Adjust min/max threshold to be below/above true min/max
The previous approach was to round to 8 decimal places; however,
this fails when both min and max are rounded up to the same value.
The solution implemented here is slightly more complex, but ensures
that the thresholds are adjusted in the correct direction (maximum
increases if it is less than the true minimum and minimum decreases
if it is greater than the true maximum).