Geometric clipping (#93)
* Test fixtures for new geometric clipping fixture

Set up to pass pdc0 for PVWatts inverter model to generate data
with clipping. 1-minute timestamp spacing can be downsampled
for tests at lower frequencies.

* Test that data with/without clipping is correctly identified

* Register the pdc0_inverter pytest.mark

pytest warns on unregistered marks to prevent typos, register this
mark so we can use it safely.

* Parametrize tests by data frequency

Test at 1, 15, 30, and 60 minute frequencies.

* Test that the correct data is flagged as clipped

Also expand parametrization to include 5 minute data.

* Down-sample data with frequency less than 10 minutes

A different method is used to calculate the clipping threshold
when data is down-sampled.

* Test with "simulated" midday cloudy period

* Use larger default window for tracking systems

* Test passing a larger window results in no clipping detected

* Test tracking=True parameter

* Don't test correctness at 1-minute frequency

Because some not-clipped data is very close to the clipping
threshold at 1-minute timestamp spacing, creating a test that
exactly captures the expected output is unreasonable. We have
tests that ensures clipping is detected at that frequency if
and only if it is present.

* Add features.clipping.geometric to api.rst

* Allow some values that are not clipped for high-frequency data

Values that are very near the clipping level cannot be reliably
distinguished from clipped data.

* Use cythonized kernel functions

Substantial performance improvements by using .transform('max')
instead of .transform(lambda xs: data[xs.index][xs].max()). Some
minor additional work is required to select the correct data before
applying .transform()

* Add test with irregular and missing data

* Raise a ValueError if ac_power is not sorted

because we are rolling over the integer indices, not time windows
the data must be sorted.

* Adjust min/max threshold to be below/above true min/max

The previous approach was to round to 8 decimal places; however,
this fails when both min and max are rounded up to the same value.
The solution implemented here is slightly more complex, but ensures
that the thresholds are adjusted in the correct direction (maximum
increases if it is less than the true minimum and minimum decreases
if it is greater than the true maximum).

Co-authored-by: Cliff Hansen <cwhanse@sandia.gov>